Code cleanup. Migrate away from C style array declarations. Remove redundant interface modifiers.  Remove unnecessary boxing/unboxing. And make use of the java diamond pattern in Java SE7.

+++ src/main/java/net/server/channel/handlers/CloseRangeDamageHandler.java
@@ -33,7 +33,6 @@ import client.MapleBuffStat;
 import client.MapleCharacter;
 import client.MapleClient;
 import client.MapleJob;
-import client.MapleStat;
 import client.Skill;
 import client.SkillFactory;
 import constants.GameConstants;
@@ -79,7 +78,7 @@ public final class CloseRangeDamageHandler extends AbstractDealDamageHandler {
         Integer comboBuff = chr.getBuffedValue(MapleBuffStat.COMBO);
         if (GameConstants.isFinisherSkill(attack.skill)) {
             if (comboBuff != null) {
-                numFinisherOrbs = comboBuff.intValue() - 1;
+                numFinisherOrbs = comboBuff - 1;
             }
             chr.handleOrbconsume();
         } else if (attack.numAttacked > 0) {
@@ -130,7 +129,7 @@ public final class CloseRangeDamageHandler extends AbstractDealDamageHandler {
             int totDamageToOneMonster = 0; // sacrifice attacks only 1 mob with 1 attack
             final Iterator<List<Integer>> dmgIt = attack.allDamage.values().iterator();
             if (dmgIt.hasNext()) {
-                totDamageToOneMonster = dmgIt.next().get(0).intValue();
+                totDamageToOneMonster = dmgIt.next().get(0);
             }
             
             chr.safeAddHP(-1 * totDamageToOneMonster * attack.getAttackEffect(chr, null).getX() / 100);
